seperated one & fivemin schedule functions
